rearrange admin functions
