+++ includes/functions.php
@@ -15,21 +15,6 @@ function wc_variation_images() {
 }
 
 /**
- * get settings options
- *
- * @param string $key Image key.
- * @param string $default_value Default value.
- * @param string $section Section Value.
- *
- * @since 1.0.0
- * @return mixed
- */
-function wc_variation_images_get_settings( $key, $default_value = '', $section = '' ) {
-	$option = get_option( $section, array() );
-	return ! empty( $option[ $key ] ) ? $option[ $key ] : $default_value;
-}
-
-/**
  * retrieve product variation image
  *
  * @param int $product_id Product ID.
@@ -78,7 +63,7 @@ function wc_variation_images_get_variation_images( $product_id, $variation_id ) 
 		$gallery_images = array_slice( $gallery_images, 0, 3 );
 	}
 
-	$hide_gallery  = wc_variation_images_get_settings( 'wc_variation_images_hide_image_slider', 'no', 'wc_variation_images_general_settings' );
+	$hide_gallery = get_option( 'wc_variation_images_hide_image_slider', 'no' );
 
 	// Add product/variation image id in gallery image array.
 	array_unshift( $gallery_images, $variation_image_id );
@@ -103,8 +88,8 @@ function wc_variation_images_get_variation_images( $product_id, $variation_id ) 
 						);
 					}
 					if ( ! empty( $attachment_id ) ) {
-						$variation_image = wc_get_gallery_image_html( $attachment_id ); // phpcs:disable WordPress.XSS.EscapeOutput.OutputNotEscaped
-						$html           .= apply_filters( 'wc_variation_images_content', $variation_image, $attachment_id ); // phpcs:disable WordPress.XSS.EscapeOutput.OutputNotEscaped
+						$variation_image = wc_get_gallery_image_html( $attachment_id );
+						$html           .= apply_filters( 'wc_variation_images_content', $variation_image, $attachment_id );
 					}
 					$flag = false;
 				}
@@ -113,7 +98,7 @@ function wc_variation_images_get_variation_images( $product_id, $variation_id ) 
 				$html .= sprintf( '<img src="%s" alt="%s" class="wp-post-image" />', esc_url( wc_placeholder_img_src( 'woocommerce_single' ) ), esc_html__( 'Awaiting product image', 'wc-variation-images' ) );
 				$html .= '</div>';
 			}
-			echo $html;
+			echo wp_kses_post( $html );
 			?>
 		</figure>
 	</div>
