+++ package.json
@@ -2,7 +2,7 @@
   "name": "wc-variation-images",
   "title": "WC Variation Images",
   "description": "Adds additional gallery images per product variation.",
-  "version": "1.1.1",
+  "version": "1.2.0",
   "homepage": "https://pluginever.com/plugins/wc-variation-images-pro/",
   "author": {
     "name": "pluginever",
