+++ phpcs.xml
@@ -8,10 +8,11 @@
 	<exclude-pattern>lib/</exclude-pattern>
 	<exclude-pattern>tests/</exclude-pattern>
 	<exclude-pattern>assets/</exclude-pattern>
-	<exclude-pattern>dist/</exclude-pattern>
+	<exclude-pattern>build/</exclude-pattern>
 	<exclude-pattern>bin/</exclude-pattern>
 	<exclude-pattern>*/node_modules/*</exclude-pattern>
 	<exclude-pattern>*/vendor/*</exclude-pattern>
+	<exclude-pattern>*/vendor-prefixed/*</exclude-pattern>
 	<exclude-pattern>*\.(css|js|xml)</exclude-pattern>
 	<arg name="basepath" value="./"/><!-- Strip the file paths down to the relevant bit -->
 	<arg name="extensions" value="php"/>
